+++ src/components/mypage/ProfileSideBar.tsx
@@ -1,8 +1,6 @@
 import { useTabContext } from '@/context/TabContext';
 import {
-  BaseProfile,
   Logout,
-  Pencil,
   ProfileAccountIcon,
   ProfileCalendarIcon,
   ProfileChecklistIcon,
@@ -13,6 +11,8 @@ import { motion } from 'framer-motion';
 import router from 'next/router';
 import React from 'react';
 
+import ProfileImage from './ProfileImage';
+
 const baseTextStyle =
   'group flex cursor-pointer items-center gap-12 rounded-xl px-20 py-8 text-lg-bold transition-all';
 const textGroupStyle =
@@ -21,10 +21,6 @@ const activeStyle = 'bg-custom-green-100 text-nomad-black';
 
 const svgStyle = 'fill-gray-400 transition-all group-hover:fill-nomad-black';
 const svgActiveStyle = 'fill-nomad-black';
-
-const whileHover = {
-  backgroundImage: 'linear-gradient(90deg, #47815b 0%, #112211 100%)',
-};
 
 type ProfileSideBarProps = {
   toggleOpen: () => void;
@@ -52,15 +48,7 @@ export default function ProfileSideBar({ toggleOpen }: ProfileSideBarProps) {
       animate={{ opacity: 1 }}
       exit={{ opacity: 0 }}
     >
-      <div className="relative">
-        <BaseProfile className="h-160 w-160" />
-        <motion.div
-          className="flex-center absolute bottom-0 right-10 h-44 w-44 cursor-pointer rounded-full bg-custom-green-200"
-          whileHover={whileHover}
-        >
-          <Pencil />
-        </motion.div>
-      </div>
+      <ProfileImage />
       <div className="flex w-full flex-col gap-12 px-12">
         <div
           className={`${baseTextStyle} ${activeTab === 'info' ? activeStyle : textGroupStyle}`}
